requirements.txt
langchain
langchain-ollama
langgraph
langchain-community
chromadb
streamlit
psycopg2-binary
python-dotenv
pymupdf
